+++ solvex/src/auth/auth.controller.ts
@@ -1,0 +1,18 @@
+import { Controller, Post, Body } from '@nestjs/common';
+import { CreateUserDto, loginDto } from 'src/users/dto/user.dto';
+import { AuthService } from './auth.service';
+
+@Controller('auth')
+export class AuthController {
+  constructor(private readonly authService: AuthService) {}
+
+  @Post('signup')
+  create(@Body() userData: CreateUserDto) {
+    return this.authService.create(userData);
+  }
+
+  @Post('signin')
+  signIn(@Body() credentials: loginDto) {
+    return this.authService.signIn(credentials);
+  }
+}
